feat(workflows): add support for arbitrary user scripts in workflows
Users can now add script steps to their workflows. The usage is quite
straightforward, just use a `script` key in your step instead of a
`command`. The script is evaluated with `bash`, and should return an
exit code of 0 if successful.

I will follow up with a 2nd PR which introduces the ability to reference
command outputs in script steps.