fix(workflows): abort if step cmd returns errors
This fixes an oversight in the initial implementation, where succeeding
steps were only aborted if the step threw an exception, but not if it
returned an error.

We now abort subsequent steps if a step command has errors in its
result, and log the error details directly in the log (since looking
into `error.log` is an undesirable extra step e.g. in a CI context).